+++ models/genre.ts
@@ -1,6 +1,7 @@
 import { Schema, model } from "mongoose";
 import singerModel from "./singer";
 import playListModel from "./playList";
+import userPlaylistModel from "./userPlaylist";
 import archiveModel from "../models/archive";
 const schema = new Schema(
   {
@@ -14,8 +15,10 @@ const schema = new Schema(
 schema.pre("deleteOne", async function (next) {
   try {
     const deletedGenre = await this.model.findOne(this.getFilter());
+    if(!deletedGenre) return next()
     await singerModel.deleteMany({ musicStyle: deletedGenre._id });
     await playListModel.deleteMany({ genre: deletedGenre._id });
+    await userPlaylistModel.deleteMany({ genre: deletedGenre._id });
     await archiveModel.deleteMany({ genre: deletedGenre._id });
     next();
   } catch (error:any) {
